Добавил класс Bar что отвечает за полоски хп/маны/и т.д. Добавил полоску ХП и маны игроку.

+++ src/main/java/com/magicode/gameplay/entity/Entity.java
@@ -7,8 +7,8 @@ public class Entity {
     protected double speed;
     protected String direction = "null";
 
-    protected double maxHealth;
-    protected double health;
+    protected int maxHealth;
+    protected int health;
     protected double regeneration;
 
     protected int collisionX;
